+++ trunk/cpsycle/driver/wasapi/wasapi.c
@@ -202,7 +202,7 @@ static int driver_open(psy_AudioDriver*);
 static int driver_close(psy_AudioDriver*);
 static int driver_dispose(psy_AudioDriver*);
 static void driver_configure(psy_AudioDriver*, psy_Properties*);
-static unsigned int samplerate(psy_AudioDriver*);
+static uintptr_t samplerate(psy_AudioDriver*);
 static const char* capturename(psy_AudioDriver*, int index);
 static int numcaptures(psy_AudioDriver*);
 static const char* playbackname(psy_AudioDriver*, int index);
@@ -502,7 +502,7 @@ void driver_configure(psy_AudioDriver* driver, psy_Properties* config)
 	}
 }
 
-unsigned int samplerate(psy_AudioDriver* self)
+uintptr_t samplerate(psy_AudioDriver* self)
 {
 	return psy_audiodriversettings_samplespersec(&((WasapiDriver*)self)->settings);
 }
